Criando pacote de persistência.
