+++ backend/src/main/java/com/auto_mendes/backend/persistence/entity/Car.java
@@ -1,4 +1,4 @@
-package com.auto_mendes.backend.entity;
+package com.auto_mendes.backend.persistence.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.PrimaryKeyJoinColumn;
@@ -9,8 +9,8 @@ import lombok.EqualsAndHashCode;
 @Data
 @EqualsAndHashCode(callSuper = true)
 @Entity
-@Table(name = "pf_tb")
+@Table(name = "car_tb")
 @PrimaryKeyJoinColumn(name = "id")
-public class PF extends Customer {
+public class Car extends Vehicle {
 	private static final long serialVersionUID = 1L;
 }
