+++ backend/src/main/java/com/auto_mendes/backend/persistence/entity/Manager.java
@@ -1,16 +1,18 @@
-package com.auto_mendes.backend.entity;
+package com.auto_mendes.backend.persistence.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.PrimaryKeyJoinColumn;
 import jakarta.persistence.Table;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
 
 @Data
 @EqualsAndHashCode(callSuper = true)
+@NoArgsConstructor
 @Entity
-@Table(name = "pj_tb")
+@Table(name = "manager_tb")
 @PrimaryKeyJoinColumn(name = "id")
-public class PJ extends Customer {
+public class Manager extends Employee {
 	private static final long serialVersionUID = 1L;
 }
