+++ backend/src/main/java/com/auto_mendes/backend/persistence/entity/Sale.java
@@ -1,10 +1,11 @@
-package com.auto_mendes.backend.entity;
+package com.auto_mendes.backend.persistence.entity;
 
 import java.io.Serializable;
 import java.math.BigDecimal;
+import java.time.LocalDateTime;
 import java.util.List;
 
-import com.auto_mendes.backend.enums.TransmissionType;
+import com.auto_mendes.backend.enums.PaymentType;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -13,8 +14,6 @@ import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.Inheritance;
-import jakarta.persistence.InheritanceType;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
@@ -27,23 +26,26 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Entity
-@Table(name = "vehicle_tb")
-@Inheritance(strategy = InheritanceType.JOINED)
-public class Vehicle implements Serializable {
+@Table(name = "sale_tb")
+public class Sale implements Serializable {
 	private static final long serialVersionUID = 1L;
-	
+
 	@Id
 	@GeneratedValue(strategy = GenerationType.UUID)
 	private String id;
-	private String plate;
+	@Column(name = "sale_date_time", nullable = false)
+	private LocalDateTime dateTime;
 	@Enumerated(EnumType.STRING)
 	@Column(nullable = false)
-	private TransmissionType transmissionType;
+	private PaymentType paymentType;
 	@Column(nullable = false, precision = 20, scale = 2)
-	private BigDecimal price;
+	private BigDecimal total;
 	@ManyToOne
-	@JoinColumn(name = "model_id")
-	private Model model;
-	@OneToMany(mappedBy = "vehicle")
+	@JoinColumn(name = "saler_id")
+	private Saler saler;
+	@ManyToOne
+	@JoinColumn(name = "customer_id")
+	private Customer customer;
+	@OneToMany(mappedBy = "sale")
 	private List<SaleVehicle> saleVehicles;
 }
