+++ backend/src/main/java/com/auto_mendes/backend/persistence/entity/SaleVehicleId.java
@@ -1,4 +1,4 @@
-package com.auto_mendes.backend.entity;
+package com.auto_mendes.backend.persistence.entity;
 
 import jakarta.persistence.Embeddable;
 import lombok.AllArgsConstructor;
