+++ backend/src/main/java/com/auto_mendes/backend/persistence/repository/AssistantManagerRepository.java
@@ -1,9 +1,9 @@
-package com.auto_mendes.backend.repository;
+package com.auto_mendes.backend.persistence.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.auto_mendes.backend.entity.AssistantManager;
+import com.auto_mendes.backend.persistence.entity.AssistantManager;
 
 @Repository
 public interface AssistantManagerRepository extends JpaRepository<AssistantManager, String> {
