+++ backend/src/main/java/com/auto_mendes/backend/validation/EmployeeValidation.java
@@ -1,8 +1,8 @@
 package com.auto_mendes.backend.validation;
 
-import com.auto_mendes.backend.entity.AssistantManager;
-import com.auto_mendes.backend.entity.Manager;
-import com.auto_mendes.backend.entity.Saler;
+import com.auto_mendes.backend.persistence.entity.AssistantManager;
+import com.auto_mendes.backend.persistence.entity.Manager;
+import com.auto_mendes.backend.persistence.entity.Saler;
 
 public interface EmployeeValidation {
 	void validadeEmployee(Manager manager);
