+++ backend/src/main/java/com/auto_mendes/backend/validation/impl/EmployeeValidationImpl.java
@@ -3,12 +3,12 @@ package com.auto_mendes.backend.validation.impl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import com.auto_mendes.backend.entity.AssistantManager;
-import com.auto_mendes.backend.entity.Manager;
-import com.auto_mendes.backend.entity.Saler;
-import com.auto_mendes.backend.repository.AssistantManagerRepository;
-import com.auto_mendes.backend.repository.ManagerRepository;
-import com.auto_mendes.backend.repository.SalerRepository;
+import com.auto_mendes.backend.persistence.entity.AssistantManager;
+import com.auto_mendes.backend.persistence.entity.Manager;
+import com.auto_mendes.backend.persistence.entity.Saler;
+import com.auto_mendes.backend.persistence.repository.AssistantManagerRepository;
+import com.auto_mendes.backend.persistence.repository.ManagerRepository;
+import com.auto_mendes.backend.persistence.repository.SalerRepository;
 import com.auto_mendes.backend.validation.EmployeeValidation;
 
 @Component
